+++ src/nanoev_udp.c
@@ -239,8 +239,8 @@ void __udp_proactor_callback(nanoev_proactor *proactor, LPOVERLAPPED overlapped)
        Internal : This member, which specifies a system-dependent status
        InternalHigh : This member, which specifies the length of the data transferred
      */
-    status = ntstatus_to_winsock_error(overlapped->Internal);
-    bytes = overlapped->InternalHigh;
+    status = ntstatus_to_winsock_error((long)overlapped->Internal);
+    bytes = (unsigned int)overlapped->InternalHigh;
 
     if (0 != status) {
         udp->flags |= NANOEV_UDP_FLAG_ERROR;
